0.2.0-beta release

+++ src/CoiniumServ/Daemon/Responses/BlockTemplateTransaction.cs
@@ -20,6 +20,9 @@
 //     license or white-label it as set out in licenses/commercial.txt.
 // 
 #endregion
+
+using System;
+
 namespace CoiniumServ.Daemon.Responses
 {
     /// <summary>
@@ -44,7 +47,7 @@ namespace CoiniumServ.Daemon.Responses
         /// difference in value between transaction inputs and outputs (in Satoshis); for coinbase transactions, this is a negative Number of the 
         /// total collected block fees (ie, not including the block subsidy); if key is not present, fee is unknown and clients MUST NOT assume there isn't one
         /// </summary>
-        public int Fee { get; set; }
+        public UInt64 Fee { get; set; }
 
         /// <summary>
         /// hash/id encoded in little-endian hexadecimal
